# openapi/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_getAllUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "example": 20,
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for email or name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by user status (active/inactive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter by user role (admin/user)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appId",
            "required": false,
            "in": "query",
            "description": "Filter by app ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns paginated users list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "pages": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users with pagination and filtering",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a user by ID",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/roles": {
      "patch": {
        "operationId": "UsersController_updateUserRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of role names"
                  }
                },
                "required": [
                  "roles"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User roles updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user roles (Admin only)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/profile": {
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User profile data to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user profile",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Doctopus API",
    "description": "Modern NestJS API with comprehensive features",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://tree2u.kobotogether.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "FileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "File ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "key": {
            "type": "string",
            "description": "File storage key/path",
            "example": "user-files/document.pdf"
          },
          "filename": {
            "type": "string",
            "description": "Original filename",
            "example": "document.pdf"
          },
          "mimetype": {
            "type": "string",
            "description": "File MIME type",
            "example": "application/pdf"
          },
          "size": {
            "type": "number",
            "description": "File size in bytes",
            "example": 1048576
          },
          "bucket": {
            "type": "string",
            "description": "Storage bucket name",
            "example": "my-app-files"
          },
          "uploadedBy": {
            "type": "string",
            "description": "User ID who uploaded the file",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "userId": {
            "type": "string",
            "description": "User ID who owns the file",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "status": {
            "type": "string",
            "description": "File status",
            "enum": [
              "PENDING",
              "UPLOADED",
              "FAILED"
            ],
            "example": "UPLOADED"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "File creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "File last update date"
          }
        },
        "required": [
          "id",
          "key",
          "filename",
          "mimetype",
          "bucket",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "displayName": {
            "type": "object",
            "description": "User display name",
            "example": "John Doe"
          },
          "avatar": {
            "type": "object",
            "description": "URL to user avatar image",
            "example": "https://example.com/avatar.jpg"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether user account is active",
            "default": true
          },
          "emailVerified": {
            "type": "boolean",
            "description": "Whether user email is verified",
            "default": false
          },
          "provider": {
            "type": "object",
            "description": "Authentication provider (local, google, github)",
            "example": "local"
          },
          "providerId": {
            "type": "object",
            "description": "ID from auth provider",
            "example": "123456789"
          },
          "roles": {
            "description": "User roles",
            "example": [
              "user"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "files": {
            "description": "Files owned by user",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Account creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Account last updated date"
          }
        },
        "required": [
          "id",
          "email",
          "isActive",
          "emailVerified",
          "roles",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "description": "User display name",
            "example": "John Doe"
          },
          "avatar": {
            "type": "string",
            "description": "URL to avatar image",
            "example": "https://example.com/avatar.jpg"
          }
        }
      },
...
    }
  }
}